requirements.txt
ffmpeg-python==0.2.0
tkinterdnd2==0.3.0
Pillow==10.0.1
opencv-python==4.8.1.78
numpy<2.0.0